requirements.txt
kafka-python==2.0.6
python-telegram-bot==20.7
google-api-python-client==2.108.0
python-dotenv==1.0.0 